[Change] separate reviews, orders, news controllers from routes
