+++ routes/ordersRoutes.js
@@ -1,78 +1,16 @@
 const express = require("express");
-const { dbPoolPromise } = require("../mySqlConnection");
 const router = express.Router();
 const { verifyJWT } = require("../JWT/jwtMiddleware");
 const { catchRequestError } = require("../utils/catchAsync");
+const controller = require("../controllers/orders_controllers");
 
-router.route("/").get(
-    verifyJWT,
-    catchRequestError(async (req, res) => {
-        const [readOrder] = await dbPoolPromise.execute(
-            "SELECT id, order_date, order_status FROM product_orders"
-        );
-        res.send(readOrder);
-    })
-);
+router.route("/").get(verifyJWT, catchRequestError(controller.findAll));
 router
     .route("/:id")
-    .post(
-        verifyJWT,
-        catchRequestError(async (req, res) => {
-            const id = req.body.id;
+    .post(verifyJWT, catchRequestError(controller.findById))
+    .put(verifyJWT, catchRequestError(controller.editById))
+    .delete(verifyJWT, catchRequestError(controller.deleteById));
 
-            const [readOrder] = await dbPoolPromise.execute(
-                "SELECT * FROM product_orders WHERE id = ?",
-                [id]
-            );
-            res.send(readOrder);
-        })
-    )
-    .put(
-        verifyJWT,
-        catchRequestError(async (req, res) => {
-            const id = req.body.id;
-            const status = req.body.status;
-
-            await dbPoolPromise.execute(
-                "UPDATE product_orders SET order_status = ? WHERE id = ?",
-                [status, id]
-            );
-
-            res.send("Order status updated");
-        })
-    )
-    .delete(
-        verifyJWT,
-        catchRequestError(async (req, res) => {
-            const id = req.body.id;
-
-            await dbPoolPromise.execute(
-                "DELETE FROM product_orders WHERE id = ?",
-                [id]
-            );
-
-            res.send("Order deleted");
-        })
-    );
-
-router.get(
-    "/status",
-    catchRequestError(async (req, res) => {
-        const [status] = await dbPoolPromise.execute(
-            "SELECT order_status FROM product_orders"
-        );
-        const validationArray = [];
-        status.map((order) => {
-            if (order.order_status === "NEW ORDER") {
-                validationArray.push(true);
-            } else {
-                validationArray.push(false);
-            }
-        });
-        const validationValue = validationArray.includes(true);
-
-        res.send(validationValue);
-    })
-);
+router.get("/status", catchRequestError(controller.status));
 
 module.exports = router;
